+++ src/fcn/weather/getCurrentWeatherData.php
@@ -191,9 +191,6 @@ class getCurrentWeatherData
     private function windDirection($windDEG)
     {
         switch ($windDEG) {
-            case (is_null($windDEG)):
-                $windDIR = 'ERROR';
-                break;
             case ($windDEG >= 11.25 && $windDEG < 33.75):
                 $windDIR = 'NNE';
                 break;